requirements.txt
absl-py==0.12.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
attrs==21.4.0
backcall==0.2.0
bleach==4.1.0
cachetools==4.2.2
certifi==2020.12.5
cffi==1.15.0
chardet==4.0.0
colorama==0.4.4
cycler==0.10.0
Cython==0.29.23
debugpy==1.5.1
decorator==5.0.9
defusedxml==0.7.1
easydict==1.9
entrypoints==0.3
filelock==3.0.12
fire==0.4.0
future==0.18.2
gdown==3.13.0
google-auth==1.30.0
google-auth-oauthlib==0.4.4
grpcio==1.37.1
idna==2.10
importlib-metadata==4.0.1
importlib-resources==5.4.0
infomap==2.3.0
ipdb==0.13.7
ipykernel==6.7.0
ipython==7.23.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.0
Jinja2==3.0.3
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client==7.1.1
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.1
llvmlite==0.36.0
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mistune==0.8.4
munch==2.5.0
nbclient==0.5.10
nbconvert==6.4.0
nbformat==5.1.3
nest-asyncio==1.5.4
notebook==6.4.7
numba==0.53.1
oauthlib==3.1.0
packaging==21.3
pandas==1.2.4
pandocfilters==1.5.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
pretrainedmodels==0.7.4
prettytable==2.1.0
prometheus-client==0.12.0
prompt-toolkit==3.0.18
protobuf==3.17.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.9.0
pynndescent==0.5.2
pyparsing==2.4.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.1
pytorch-ignite==0.1.2
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.0
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
Send2Trash==1.8.0
sklearn==0.0
tabulate==0.8.9
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.2
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
toml==0.10.2




